[0821] Fixed a few issues in Zul'Farrak
* Fixed a few issues in Zul'Farrak. Thanks Nekoi for reporting.
* Added missing loot entry for creature 8876 (Sandfury Acolyte)
* Fixed quest Divino-matic rod
* Added waypoint movement to creature 7795 (Hydromancer Velratha) in
Zul'Farrak. She will now move around the sacred pool.
Source: YTDB
Fixed walk speed for creature 7273 (Gahz'rilla) in Zul'Farrak because
it was moving way to quickly
Fixed addon for creature 5648 (Sandfury Shadowcaster) because they were
using an aura addon and an EventAI script to add the same spell 20798
(Demon Skin) resulting in the spell being self canceled
This closes #175 and #174