[0971] Added dedicated development directory
* Created a new 'dev' directory at repository root. This will allow dev
to:
- have development update files applied after all released updates
(including core, SD2 and ACID). This will reduce conflicts and prevent
development updates to be erased by regular updates. This also allows
development to be done on an automatically up-to-date and clean
repository
- keep the dev files separated from the regular update files, making easier to
manage them, especially when updates directory has near a hundred files
in it

The only naming convention for dev files is to have a `.sql` extension

In order to use the dev directory, set the `DEV_UPDATES` variable to
`YES` in the `InstallFullDB.config` file.

See `dev/READ ME` for more info.