[1361] Correct data for some ranged weapons
* Correct data for some ranged weapons
This closes #858
http://wow.gamepedia.com/Patch_1.12.0
Ranged Elemental Damage: Our ranged combat system does not allow a
ranged weapon to do a mix of Elemental (Fire, Frost, Arcane, etc.) and
Physical damage. Many ranged weapons existed
which were listed as doing Elemental damage and did not function
properly (the damage was dealt, but was treated as Physical). All of
those weapons have been changed to deliver the
Elemental damage as a chance on hit effect. [Hurricane] was previously
changed this way in 1.11. The following ranged weapons are also fixed in
1.12:

Signed-off-by: Cala <calaftp@free.fr>